Re-introducing --no-default-features behavior
- `check-passthrough` must be enabled on the sys dependency to even allow setting the old `--no-defaults` behavior
- `dont-use-deprecated` feature of the sys package allows suppression of old `--no-defaults` to be set as a default in the rs package
- `dont-use-deprecated` must be **disabled** by passing `--no-defaults` to rs package to enable the old `--no-defaults` behavior in sys

The old use of --no-default-features applies to shaderc-rs, but now it needs to be passed through to shaderc-sys.  This is done by passing an inverted flag to suppress the behavior by default from shaderc-rs.  When the suppression is removed with --no-default-features, the old behavior is re-enabled and shaderc-sys looks for a libshaderc_combined.a in the out directory

When building shaderc-sys directly, only check for the passthrough flag if it could be enabled.  It's a package opt-in deprecated behavior.  It definitely needs deprecation.