Require ninja when building with MSVC.
By default shaderc is built with msbuild when building with MSVC.
Unfortunately this often causes the build to fail due to excessively
long path names created by the msbuild project generator. This is
especially true when shaderc is built as a dependency of another
project, since in that case it will build in a nested subdirectory.

The best solution to this problem is to build with ninja instead, but
this solution is not obvious especially for users of downstream
projects, who are unlikely to have read the shaderc README.

To avoid this problem, we make ninja a hard requirement when building
with MSVC. This causes the build to fail with 'couldn't find required
command: "ninja"' when ninja is not available. This at least points the
user in the right direction.