SHADERC_STATIC="./here/"
If you are here because your windows or macos builds are breaking, I apologize.  The changes were not developed on these platforms.  You will have to research older versions of the crate to establish whether they were in a "just works" state and if not, research your relevant platform options and make it work.

This commit implements several new build behaviors:

- Building libshaderc from source is now fallback behavior on Linux
- Linux will check "/usr/lib/" for libraries by default
- SPIRV and glslang libraries _may_ be picked up correctly if they exist in the same location as libshaderc, but funky combinations could break
- Specification of a static library via SHADERC_STATIC, SHADERC_LIB_DIR, or SHADERC_LIB_PATH instead of the hacky --no-default-options
- tests will build correctly when using hacky --no-default-options paths (stdc was not being included)