Create shaderc-sys for FFI, building, and linking (#46)
This commit creates a new shaderc-sys crate for Shaderc library FFI, building and linking.
This change separates the concerns and is the recommended way of handling
low-level C libraries.

This commit also enhances Shaderc-sys with the ability to detect and use pre-built
Shaderc libraries. It has the following building and linking behavior:

1. The option `--features build-from-source` will prevent automatic library
   detection and force building from source.
2. If the `SHADERC_LIB_DIR` environment variable is set to
   `/path/to/shaderc/libs/`, it will take precedence and `libshaderc_combined.a`
   (and the glsang and SPIRV libraries on Linux) will be searched in the
   `/path/to/shaderc/libs/` directory.
3. On Linux, `/usr/lib/` will be automatically searched for system libraries
   if none of the above were given.
4. If no other option was set or succeeded, shaderc-sys will fall back to
   checking out and compiling a copy of shaderc.  This procedure is quite slow.

This commit also bumps version to 0.5.0. Version 0.4.0 has been skipped and tag will not
be used.  History for a prospective split at 0.4.0 into two repos has been obliterated.